LayoutTests/fast/parser/pre-html5-parser-quirks.html

<script>
    if (window.layoutTestController) {
        layoutTestController.overridePreference("WebKitUsePreHTML5ParserQuirks", "1");
        layoutTestController.dumpAsText();
        layoutTestController.dumpChildFramesAsText();
    }
</script>

<p>This tests that certain pre-HTML5 parser rules are applied when the WebKitUsePreHTML5ParserQuirks preference is set. See <a href="https://webkit.org/b/40961">https://webkit.org/b/40961</a> and <a href="https://webkit.org/b/42909">https://webkit.org/b/42909</a> for details.</p>

<iframe src="resources/pre-html5-parser-quirk-after-attribute-name-state.html"></iframe>
<iframe src="resources/pre-html5-parser-quirk-after-attribute-value-quoted-state.html"></iframe>
<iframe src="resources/pre-html5-parser-quirk-after-attribute-value-unquoted-state.html"></iframe>
<iframe src="resources/pre-html5-parser-quirk-attribute-name-state.html"></iframe>
<iframe src="resources/pre-html5-parser-quirk-before-attribute-name-state.html"></iframe>
<iframe src="resources/pre-html5-parser-quirk-document-fragment.html"></iframe>
<iframe src="resources/pre-html5-parser-quirk-tag-name-state.html"></iframe>
<iframe src="resources/pre-html5-parser-quirk-self-closing-script-in-head.html"></iframe>
<iframe src="resources/pre-html5-parser-quirk-self-closing-script-in-body.html"></iframe>
